move secret helpers to common package
