+++ internal/generator/fluentd/output/fluentdforward/cert-key.go
@@ -1,7 +1,7 @@
 package fluentdforward
 
 import (
-	"github.com/openshift/cluster-logging-operator/internal/generator/fluentd/output/security"
+	"github.com/openshift/cluster-logging-operator/internal/generator/helpers/security"
 )
 
 type TLSCertKey security.TLSCertKey
