+++ internal/generator/fluentd/output/http/cert-key.go
@@ -1,7 +1,7 @@
 package http
 
 import (
-	"github.com/openshift/cluster-logging-operator/internal/generator/fluentd/output/security"
+	"github.com/openshift/cluster-logging-operator/internal/generator/helpers/security"
 )
 
 type TLSKeyCert security.TLSCertKey
