+++ internal/generator/fluentd/output/http/passphrase.go
@@ -1,6 +1,8 @@
 package http
 
-import "github.com/openshift/cluster-logging-operator/internal/generator/fluentd/output/security"
+import (
+	"github.com/openshift/cluster-logging-operator/internal/generator/helpers/security"
+)
 
 type Passphrase security.Passphrase
