+++ internal/generator/fluentd/output/http/userpass.go
@@ -1,7 +1,7 @@
 package http
 
 import (
-	security2 "github.com/openshift/cluster-logging-operator/internal/generator/fluentd/output/security"
+	security2 "github.com/openshift/cluster-logging-operator/internal/generator/helpers/security"
 )
 
 type UserNamePass security2.UserNamePass
