+++ internal/generator/fluentd/output/kafka/ca-file.go
@@ -1,7 +1,7 @@
 package kafka
 
 import (
-	"github.com/openshift/cluster-logging-operator/internal/generator/fluentd/output/security"
+	"github.com/openshift/cluster-logging-operator/internal/generator/helpers/security"
 )
 
 type CAFile security.CAFile
