+++ internal/generator/fluentd/output/kafka/cert-key.go
@@ -1,7 +1,7 @@
 package kafka
 
 import (
-	"github.com/openshift/cluster-logging-operator/internal/generator/fluentd/output/security"
+	"github.com/openshift/cluster-logging-operator/internal/generator/helpers/security"
 )
 
 type TLSKeyCert security.TLSCertKey
