+++ internal/generator/fluentd/output/loki/bearer-token-file.go
@@ -1,7 +1,7 @@
 package loki
 
 import (
-	"github.com/openshift/cluster-logging-operator/internal/generator/fluentd/output/security"
+	"github.com/openshift/cluster-logging-operator/internal/generator/helpers/security"
 )
 
 type BearerTokenFile security.BearerTokenFile
