+++ internal/generator/fluentd/output/syslog/syslog.go
@@ -2,6 +2,7 @@ package syslog
 
 import (
 	"fmt"
+	"github.com/openshift/cluster-logging-operator/internal/generator/helpers/security"
 	"regexp"
 	"strings"
 
@@ -13,7 +14,6 @@ import (
 	"github.com/openshift/cluster-logging-operator/internal/generator/fluentd/helpers"
 	"github.com/openshift/cluster-logging-operator/internal/generator/fluentd/normalize"
 	"github.com/openshift/cluster-logging-operator/internal/generator/fluentd/output"
-	"github.com/openshift/cluster-logging-operator/internal/generator/fluentd/output/security"
 	"github.com/openshift/cluster-logging-operator/internal/generator/fluentd/source"
 	genhelper "github.com/openshift/cluster-logging-operator/internal/generator/helpers"
 	urlhelper "github.com/openshift/cluster-logging-operator/internal/generator/url"
