+++ internal/generator/helpers/security/security.go
