+++ internal/generator/helpers/security/security_test.go
