+++ internal/generator/vector/output/kafka/kafka_test.go
@@ -1,17 +1,16 @@
 package kafka
 
 import (
+	"github.com/openshift/cluster-logging-operator/internal/generator/helpers/security"
 	"testing"
 
 	"github.com/openshift/cluster-logging-operator/test/helpers"
-
-	"github.com/openshift/cluster-logging-operator/internal/generator"
-	"github.com/openshift/cluster-logging-operator/internal/generator/fluentd/output/security"
 
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/ginkgo/extensions/table"
 	. "github.com/onsi/gomega"
 	logging "github.com/openshift/cluster-logging-operator/apis/logging/v1"
+	"github.com/openshift/cluster-logging-operator/internal/generator"
 	corev1 "k8s.io/api/core/v1"
 )
